Refine recursion test
Add hard limit to self calling recursion. It seems the message for a given timer will always arrive regardless of whether that timer is killed - that's bad as a timer could be due far in the future, and so flushing the message queue wouldn't help. However making then immediately destroying seems to work, with an application.wait.